Extract transcoding options to its own hook
